restructure seed files
