+++ seeds/load-all-data.js
@@ -1,7 +1,13 @@
 const fs = require('fs');
 
 exports.seed = async function (knex) {
-  
+
+  await knex("Playlist").del();
+  await knex("Song").del();
+  await knex("ShowDJs").del();
+  await knex("Show").del();
+  await knex("DJs").del(); 
+
   const contentsShow = fs.readFileSync('./data/shows.json');
   const shows = JSON.parse(contentsShow);
 
@@ -16,12 +22,10 @@ exports.seed = async function (knex) {
       id: show.id
     })
   );
-  await knex("Show").del();
   await knex.batchInsert("Show", showData, 100);
 
   const djContents = fs.readFileSync("./data/djs.json");
   const djData = JSON.parse(djContents);
-  await knex("DJs").del();
   await knex.batchInsert("DJs", djData, 100);
 
   const djMap = [];
@@ -30,6 +34,13 @@ exports.seed = async function (knex) {
       djMap.push({showId: show.id, djId: id});
     });
   });
-  await knex("ShowDJs").del();
   await knex.batchInsert("ShowDJs", djMap, 100);
+
+  const playlists = fs.readFileSync("./data/playlists.json");
+  const playlistData = JSON.parse(playlists);
+  await knex.batchInsert("Playlist", playlistData, 100);
+
+  const songs = fs.readFileSync("./data/songs.json");
+  const songData = JSON.parse(songs);
+  await knex.batchInsert("Song", songData, 100);
 };
